requirements.txt
APScheduler==3.10.4
asgiref==3.7.2
certifi==2023.7.22
cffi==1.17.1
charset-normalizer==3.2.0
Django==5.0
django-lifecycle==1.1.2
h3==4.2.2
idna==3.4
numpy==2.2.6
pycparser==2.22
pyTelegramBotAPI==4.13.0
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
setuptools==65.5.0
six==1.16.0
sqlparse==0.4.4
timezonefinder==6.5.9
tzdata==2023.3
tzlocal==5.0.1
urllib3==2.0.4
